Reworked menu from last lab

+++ motors.h
@@ -1,4 +1,4 @@
 int COUNTS_PER_ROTATION = 64;
 
-void drive_motor(float speed);
+void drive_motor(int speed);
 void initialize_motor();
